pyproject.toml
[project]
name = "autograph-concentriccircles"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "matplotlib>=3.10.8",
    "opencv-contrib-python>=4.13.0.92",
    "pillow>=12.1.1",
    "pytest>=9.0.2",
    "scipy>=1.17.1",
    "svgwrite>=1.4.3",
]
